fix(tmux): split acquireFlockLock into platform files for Windows build
syscall.Flock is not available on Windows, causing build failures.
Move acquireFlockLock to flock_unix.go (!windows) and add a
flock_windows.go stub using sync.Mutex.TryLock for in-process locking.
tmux is not supported on Windows so the stub only needs to satisfy
the compiler; tests that invoke it will work correctly.


Executed-By: gastown/crew/zhora
Rig: gastown
Role: crew
